requirements.txt
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.1
PyYAML==6.0
redis==4.0.2
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
matplotlib==3.5.3
seaborn==0.11.2 